ALYtools: t_dependent_Nuclei_ratio_FRET - bug fixing

+++ Classes/@ALYtools_data_controller/do_t_dependent_Nuclei_ratio_FRET_Segmentation.m
@@ -84,7 +84,7 @@
                k 
             end
             %
-            sgm = uint8(icyvol);
+            sgm = uint16(icyvol);
                                                            
                 if send_to_Icy                
                     try
